+++ DelvUI/changelog.md
@@ -1,6 +1,10 @@
 # 0.6.3.3
+Features:
+- Added element background colors for Black Mage's custom mana bar.
+
 Fixes:
 - GCD Indicator now works for melee classes under level 30.
+- Fixed Black Mage's mana bar "Use Element Color" setting not disabling properly.
 
 # 0.6.3.2
 Fixes: